requirements.txt
fastapi==0.61.1
tortoise-orm==0.16.17
pydantic==1.7.2